employee-analytics/requirements/base.txt

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Faker==19.6.2
Pillow==10.1.0
